requirements.txt
# Updated requirements for SoccerPredictor with API integration
requests==2.31.0
sqlite3-utils==3.35.2
python-dotenv==1.0.0

# Core ML and Data Processing
absl-py==2.0.0
pandas==2.3.1
numpy==2.3.1
scikit-learn==1.7.1
matplotlib==3.10.3
PyYAML==6.0.2

# Web Framework (updated versions)
dash==3.1.1
plotly==6.2.0
Flask==3.1.1
Flask-Compress==2.0.0
Flask-CORS==4.0.0

# Deep Learning
tensorflow==2.18.0
Keras==3.8.0

# Utilities
click==8.1.7
python-dateutil==2.9.0
pytz==2025.1
joblib==1.4.2

# API and HTTP clients
httpx==0.26.0
aiohttp==3.9.1

# Data validation and security
pydantic==2.10.2
marshmallow==3.20.1

# Development and testing
pytest==8.3.4
black==24.12.0
flake8==7.0.0

# Production dependencies
gunicorn==21.2.0
python-json-logger==2.0.7
